tjdecomp: Fix int overflow with large dimensions
In the width * height * tjPixelSize[pixelFormat] * sampleSize
computation, the width * height intermediate was computed using signed
integer arithmetic, which overflowed and wrapped around if the result
was greater than INT_MAX (2,147,483,647).  With certain values of width
and height, this caused the caller-supplied destination buffer to be
smaller than the library expected, leading to a buffer overrun and
subsequent segfault.

This issue was not a security vulnerability, since it was wholly
contained within the tjdecomp example program, tjdecomp is distributed
only as source code (it is part of the documentation, not the library),
and it is not exposed to arbitrary input data.  The worst case arising
from the issue would have been if a security-critical downstream program
that handles arbitrary input data adopted the errant code, thus
introducing a security vulnerability into said downstream program.
However, that is a highly unlikely scenario, and even if it occurred, it
would not be a security vulnerability in libjpeg-turbo itself.  (As my
mother used to say, just because someone tells you to jump off a bridge
doesn't mean you have to do it.)  Otherwise, the worst case would have
been if a user attempted to feed a malformed image into tjdecomp and it
segfaulted, but again, the segfault would have been wholly contained
within the tjdecomp example program.

Essentially, tjdecomp was abusing the TurboJPEG API.

+++ src/tjdecomp.c
@@ -342,9 +342,10 @@ int main(int argc, char **argv)
       sampleSize > (unsigned long long)((size_t)-1))
     THROW("allocating uncompressed image buffer", "Image is too large");
 #endif
+
   if ((dstBuf =
-       (unsigned char *)malloc(width * height * tjPixelSize[pixelFormat] *
-                               sampleSize)) == NULL)
+       (unsigned char *)malloc((size_t)width * height *
+                               tjPixelSize[pixelFormat] * sampleSize)) == NULL)
     THROW_UNIX("allocating uncompressed image buffer");
 
   if (precision <= 8) {
